refactor: Update ReviewService
Add validations.. for writing & delete

+++ src/main/java/com/mip/sharebnb/service/ReviewService.java
@@ -12,11 +12,13 @@ import com.mip.sharebnb.repository.ReservationRepository;
 import com.mip.sharebnb.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 
 @Service
 @RequiredArgsConstructor
+@Transactional
 public class ReviewService {
 
     private final ReviewRepository reviewRepository;
@@ -81,6 +83,32 @@ public class ReviewService {
                 .findReviewByReservationId(reviewDto.getReservationId())
                 .orElseThrow(() -> new DataNotFoundException("Review Not Found"));
 
+        Reservation reservation = reservationRepository.findById(reviewDto.getReservationId())
+                .orElseThrow(() -> new DataNotFoundException("Reservation Not Found"));
+
+        Accommodation accommodation = reservation.getAccommodation();
+        if (accommodation == null) {
+            throw new DataNotFoundException("Accommodation Not Found");
+        }
+
+        if (accommodation.getId() != reviewDto.getAccommodationId()) {
+            throw new InvalidInputException("Accommodation Not Matched");
+        }
+
+        Member member = reservation.getMember();
+        if (member == null) {
+            throw new DataNotFoundException("Member Not Found");
+        }
+
+        if (member.getId() != reviewDto.getMemberId()) {
+            throw new InvalidInputException("Member Not Matched");
+        }
+
+        float originRating = accommodation.getRating();
+        float newRating = originRating + (reviewDto.getRating() - originReview.getRating()) / accommodation.getReviewNum();
+        accommodation.setRating(newRating);
+
+        originReview.setAccommodation(accommodation);
         originReview.setContent(reviewDto.getContent());
         originReview.setRating(reviewDto.getRating());
         originReview.setCreatedDate(LocalDate.now());
@@ -89,10 +117,25 @@ public class ReviewService {
     }
 
     public void deleteReview(Long id) {
-        Review originReview = reviewRepository.findById(id)
+        Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("Review Not Found"));
 
-        originReview.getReservation().setIsWrittenReview(false);
+        Reservation reservation = review.getReservation();
+        if (reservation == null) {
+            throw new DataNotFoundException("Reservation Not Found");
+        }
+        reservation.setIsWrittenReview(false);
+
+
+        Accommodation accommodation = reservation.getAccommodation();
+        if (accommodation == null) {
+            throw new DataNotFoundException("Accommodation Not Found");
+        }
+
+        accommodation.setRating((accommodation.getRating() * accommodation.getReviewNum() - review.getRating())
+                / (accommodation.getReviewNum() - 1));
+        accommodation.setReviewNum(accommodation.getReviewNum() - 1);
+        accommodation.setHostReviewNum(accommodation.getHostReviewNum() - 1);
 
         reviewRepository.deleteById(id);
     }
